Refactor navigation structure and add RegisterScreen component
