+++ App.js
@@ -2,9 +2,10 @@ import React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import { SafeAreaView } from 'react-native';
-import SplashScreen from './src/feature/Screens/WelcomeScreen';
-import HomeScreen from './screens/homescreen'; // Add import for HomeScreen
+import WelcomeScreen from './src/feature/Screens/WelcomeScreen';
+import HomeScreen from './src/feature/Screens/homescreen'; // Add import for HomeScreen
 import LoginScreen from './src/feature/Screens/LoginScreen';
+import RegisterScreen from './src/feature/Screens/RegisterScreen';
 
 const Stack = createNativeStackNavigator();
 
@@ -12,10 +13,11 @@ export default function App() {
   return (
     <NavigationContainer>
       <SafeAreaView style={{ flex: 1 }}>
-        <Stack.Navigator initialRouteName="Splash">
-          <Stack.Screen name="Splash" component={SplashScreen} options={{ headerShown: false }} />
+        <Stack.Navigator initialRouteName="Welcome">
+          <Stack.Screen name="Welcome" component={WelcomeScreen} options={{ headerShown: false }} />
           <Stack.Screen name="Login" component={LoginScreen} options={{ headerShown: false }} />
           <Stack.Screen name="Home" component={HomeScreen} options={{ headerShown: false }} />
+          <Stack.Screen name="Register" component={RegisterScreen} options={{ headerShown: false }} />
         </Stack.Navigator>
       </SafeAreaView>
     </NavigationContainer>
